compatibility fix

+++ analysis/src/WZsynchro.cc
@@ -1235,7 +1235,7 @@ WZsynchro::getFR(Candidate* cand, int idx) {
 
   int wp=std::abs(cand->pdgId()==11)?SusyModule::kTight:SusyModule::kMedium;
 
-  if(_FR.find("C")!=string::npos) ptVal=std::max(_wzMod->conePt(idx,wp), (float)ptM);
+  if(_FR.find("C")!=string::npos) ptVal=std::max(_wzMod->conePt(idx,wp), (double)ptM);
   if(_FR.find("J")!=string::npos) ptVal/=_vc->get("LepGood_jetPtRatiov2", idx);
 
   ptVal=std::max(ptVal, ptM);
